+++ backend/canisters/identity/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Add ability to debug user principal migrations ([#5750](https://github.com/open-chat-labs/open-chat/pull/5750))
+
 ## [[2.0.1155](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1155-identity)] - 2024-04-28
 
 ### Changed